feat: extra resolveconfig from api-impl
