+++ src/cli.ts
@@ -1,13 +1,13 @@
-import { resolveUserConfig, build } from './api-impl'
+import { build } from './api-impl'
 import { print } from './common/logger'
+import { resolveConfig } from './cli-impl'
 
 export default (async function () {
-  const options = {}
   try {
-    Object.assign(options, await resolveUserConfig())
+    const options = await resolveConfig()
+    await build(options)
   } catch (error) {
     if (error instanceof Error) print.danger(error.message)
     process.exit(1)
   }
-  await build(options)
 })()
