+++ src/common/fs.ts
@@ -11,3 +11,12 @@ export const readJson = (path: string) =>
     .readFile(path, 'utf-8')
     .then((r) => JSON.parse(r))
     .catch((e) => e)
+
+export const existSync = (path: string) => {
+  try {
+    fs.accessSync(path, fs.constants.F_OK)
+    return true
+  } catch (error) {
+    return false
+  }
+}
