+++ src/common/utils.ts
@@ -1,7 +1,6 @@
 export const serialize = <T>(source: Record<string, T>): T[] => Object.values(source)
 
-// If the user does not have too many configuration options, the yaml file can be used
-export const mayBeConfig = ['bump.config.js', 'bump.config.ts', 'bump.yml', 'bump.yaml']
+export const configFiles = ['bump.config.js', 'bump.config.ts']
 
 const BASIC_TYPES = {
   Null: 'Null',
@@ -39,3 +38,5 @@ export const loadModule = (alias: string) => {
     delete require.cache[require.resolve(alias)]
   }
 }
+
+export const len = (tar: string | unknown[]) => tar.length
